Mongobackend/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
pydantic>=1.8.2
python-dotenv>=0.19.0
pymongo>=4.3.3
scikit-learn>=1.2.2  # For ML model
joblib>=1.3.2        # For model loading
psycopg2-binary>=2.9.3  # For SQL
tenacity>=8.2.2      # For retries